refactor:servicesにあったcheckerをdomain/warningsに昇格
