+++ app/domain/warnings/if_condition_warning_checker.rb
@@ -1,0 +1,98 @@
+module Warnings
+  class IfConditionWarningChecker
+  WARNING_KEYWORDS = {
+    # 恒常・頻度系
+    "常に" => "『常に』は例外を含みやすく、実行条件として曖昧になりがちです。",
+    "つねに" => "『常に』は例外を含みやすく、実行条件として曖昧になりがちです。",
+
+    "いつも" => "『いつも』は具体的なタイミングや状況が不明確です。",
+    "何時も" => "『いつも』は具体的なタイミングや状況が不明確です。",
+
+    "毎回" => "『毎回』は発生条件が特定できず、行動に結びつきにくくなります。",
+    "まいかい" => "『毎回』は発生条件が特定できず、行動に結びつきにくくなります。",
+
+    # 時間未定系
+    "後で" => "『後で／あとで』は実行時点が定義されておらず、先延ばしを招きやすい表現です。",
+    "あとで" => "『後で／あとで』は実行時点が定義されておらず、先延ばしを招きやすい表現です。",
+
+    "そのうち" => "『そのうち』は発生タイミングが不明で、実行されない可能性が高いです。",
+    "その内"   => "『そのうち』は発生タイミングが不明で、実行されない可能性が高いです。",
+
+    "いずれ" => "『いずれ』は具体的な条件や期限がなく、行動を先送りしやすくなります。",
+    "何れ"   => "『いずれ』は具体的な条件や期限がなく、行動を先送りしやすくなります。",
+
+    # 程度・努力量依存系
+    "できるだけ" => "『できるだけ／なるべく』は達成基準が曖昧で、判断に迷いやすくなります。",
+    "なるべく"   => "『できるだけ／なるべく』は達成基準が曖昧で、判断に迷いやすくなります。",
+
+    "可能なら"   => "『可能なら』は実行が任意になり、習慣化しにくくなります。",
+    "可能であれば" => "『可能なら』は実行が任意になり、習慣化しにくくなります。",
+
+    # 感情・状態依存系
+    "気が向いたら" => "『気が向いたら』は感情依存になり、再現性が低くなります。",
+    "きがむいたら" => "『気が向いたら』は感情依存になり、再現性が低くなります。",
+
+    "やる気が出たら" => "『やる気が出たら』は内的状態に左右され、行動のきっかけとして不安定です。",
+    "やるきがでたら" => "『やる気が出たら』は内的状態に左右され、行動のきっかけとして不安定です。",
+
+    "余裕があれば" => "『余裕があれば』は状況判断が必要になり、実行が後回しになります。",
+    "よゆうがあれば" => "『余裕があれば』は状況判断が必要になり、実行が後回しになります。",
+
+    "落ち着いたら" => "『落ち着いたら』は状態変化が曖昧で、開始条件として弱い表現です。",
+    "おちついたら" => "『落ち着いたら』は状態変化が曖昧で、開始条件として弱い表現です。",
+
+    # 主観判断・記憶依存系
+    "暇なとき" => "『暇なとき』は主観的な状態に依存し、実行タイミングが定まりません。",
+    "ひまなとき" => "『暇なとき』は主観的な状態に依存し、実行タイミングが定まりません。",
+
+    "時間があれば" => "『時間があれば』は判断コストが高く、行動が発生しにくくなります。",
+    "じかんがあれば" => "『時間があれば』は判断コストが高く、行動が発生しにくくなります。",
+
+    "思い出したら" => "『思い出したら』は外部トリガーがなく、忘却されやすい条件です。",
+    "おもいだしたら" => "『思い出したら』は外部トリガーがなく、忘却されやすい条件です。",
+
+    "忘れなければ" => "『忘れなければ』は記憶依存で、再現性が低くなります。",
+    "わすれなければ" => "『忘れなければ』は記憶依存で、再現性が低くなります。",
+
+    "できたら" => "『できたら』は成功条件が不明確で、行動に結びつきにくくなります。",
+
+    # 判断先送り系
+    "必要になったら" => "『必要になったら』は判断基準が曖昧で、開始が遅れがちです。",
+    "ひつようになったら" => "『必要になったら』は判断基準が曖昧で、開始が遅れがちです。",
+
+    "タイミングを見て" => "『タイミングを見て』は主観判断が入り、実行が不安定になります。",
+    "様子を見て" => "『様子を見て』は行動開始の条件が定義されていません。"
+  }.freeze
+
+    # def self.check(check_text)
+    #   return [] if check_text.blank?
+
+    #   WARNING_KEYWORDS.filter_map do |keyword, message|
+    #     { field: :if_condition, message: message } if check_text.include?(keyword)
+    #   end
+    # end
+
+    def self.check(text)
+        return [] if text.blank?
+
+        warnings = []
+
+        concept_key = :ambiguous_trigger_expression
+        concept = WarningConcepts::AmbiguousTriggerExpression.definition
+
+        concept[:patterns].each do |pattern_key, pattern|
+          matched = pattern[:matchers].select { |w| text.include?(w) }
+          next if matched.empty?
+
+          warnings << {
+            field: :if_condition,
+            concept: concept_key,
+            pattern: pattern_key,
+            matches: matched
+          }
+        end
+
+        warnings
+      end
+  end
+end
