+++ app/forms/if_then_rule_form.rb
@@ -66,7 +66,7 @@ class IfThenRuleForm
   def collect_warnings
     # 編集時と新規作成で挙動が変化するものは編集前の元のルールである@if_then_rule_of_modelを渡す。
     @warnings = []
-    @warnings += ::IfConditionWarningChecker.check(if_condition)
+    @warnings += Warnings::IfConditionWarningChecker.check(if_condition)
     @warnings += ::IfConditionDuplicateChecker.check(
       user: @current_user,
       if_condition: if_condition,
